Add package game and move functions to better fitting files
