+++ src/game/success_test.py
@@ -1,6 +1,5 @@
 import unittest
-
-from game import check_success
+from game.success import check_success
 
 
 class TestGameMethods(unittest.TestCase):
